resolve_ids_found_in_events: Delete unused resource_types and instances edgetypes() methods. Refactor to fix pylint errors. Replace resource graph lazy evaluation with having the update_graph task always unpack into memory. Change None to Unknown.  Added instance_, user_, and tenant_ _name and _type keys to the /core/events/search/ results.
