+++ goldstone/core/tasks.py
@@ -42,8 +42,8 @@ def delete_indices(prefix,
 
 @celery_app.task()
 def update_graph():
-    """Update the Resource graph's persistent data from the current OpenStack
-    cloud state.
+    """Update the Resource graph's persistent and in-memory data from the
+    current OpenStack cloud state.
 
     Nodes are:
        - deleted if they are no longer in the OpenStack cloud.
@@ -51,11 +51,16 @@ def update_graph():
        - updated from the cloud if they are already in the graph.
 
     """
+    from goldstone.core import resource
     from goldstone.glance.utils import update_glance_nodes
     from goldstone.nova.utils import update_nova_nodes
     from goldstone.cinder.utils import update_cinder_nodes
 
+    # Update the persistent state
     for services in [update_glance_nodes,
                      update_nova_nodes,
                      update_cinder_nodes]:
         services()
+
+    # Update the in-memory state
+    resource.instances.unpack()
